+++ packages/react-magma-dom/src/components/DatePicker/CalendarContext.ts
@@ -17,6 +17,8 @@ export interface CalendarContextInterface {
   onPrevMonthClick: () => void;
   onNextMonthClick: () => void;
   setDateFocused: (value: boolean) => void;
+  setFocusedDate: (day: Date) => void;
+  setFocusedTodayDate: (event: React.SyntheticEvent) => void;
 }
 
 export const CalendarContext = React.createContext<CalendarContextInterface>({
@@ -33,4 +35,6 @@ export const CalendarContext = React.createContext<CalendarContextInterface>({
   onPrevMonthClick: () => {},
   onNextMonthClick: () => {},
   setDateFocused: (value: boolean) => {},
+  setFocusedDate: (day: Date) => {},
+  setFocusedTodayDate: (event: React.SyntheticEvent) => {},
 });