+++ packages/react-magma-dom/src/i18n/interface.ts
@@ -120,6 +120,9 @@ export interface I18nInterface {
     disabledDayAriaLabel: string;
     selectedDayAriaLabel: string;
     todayAriaLabel: string;
+    today: string;
+    navigateToCurrentDateAriaLabel: string;
+    backToCalendarAriaLabel: string;
 
     helpModal: {
       header: string;