+++ website/react-magma-docs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 5.2.0-next.12
+
+### Patch Changes
+
+- 685b29014: fix (Date Picker): Update focus behaviour - shifting focus to selected date.
+
+## 5.2.0-next.11
+
+### Patch Changes
+
+- 8887aeeba: chore: fix instructions and update message for `LineChart`
+- Updated dependencies [6893e3223]
+  - react-magma-dom@4.10.0-next.18
+
 ## 5.2.0-next.10
 
 ### Minor Changes